updated product converter
