+++ tools/actions/convert/test/fixtures/product-converted.html
@@ -26,12 +26,8 @@
       </p>
       <div class="section-metadata">
         <div>
-          <div>icon</div>
+          <div>tabIcon</div>
           <div>/content/dam/danaher/system/icons/Bookmark.svg</div>
-        </div>
-        <div>
-          <div>tabId</div>
-          <div>overview</div>
         </div>
         <div>
           <div>tabName</div>
@@ -50,12 +46,8 @@
       </p>
       <div class="section-metadata">
         <div>
-          <div>icon</div>
+          <div>tabIcon</div>
           <div>/content/dam/danaher/system/icons/Chart bar.svg</div>
-        </div>
-        <div>
-          <div>tabId</div>
-          <div>specification</div>
         </div>
         <div>
           <div>tabName</div>
@@ -65,12 +57,8 @@
       <hr>
       <div class="section-metadata">
         <div>
-          <div>icon</div>
+          <div>tabIcon</div>
           <div>/content/dam/danaher/system/icons/Cube.svg</div>
-        </div>
-        <div>
-          <div>tabId</div>
-          <div>products</div>
         </div>
         <div>
           <div>tabName</div>
@@ -81,12 +69,8 @@
       <p>Resource page from Coveo atomic template</p>
       <div class="section-metadata">
         <div>
-          <div>icon</div>
+          <div>tabIcon</div>
           <div>/content/dam/danaher/system/icons/Collection.svg</div>
-        </div>
-        <div>
-          <div>tabId</div>
-          <div>resources</div>
         </div>
         <div>
           <div>tabName</div>
@@ -99,12 +83,8 @@
       </p>
       <div class="section-metadata">
         <div>
-          <div>icon</div>
+          <div>tabIcon</div>
           <div>/content/dam/danaher/system/icons/Clipboard list.svg</div>
-        </div>
-        <div>
-          <div>tabId</div>
-          <div>citations</div>
         </div>
         <div>
           <div>tabName</div>
