+++ tools/importer/transformers/productMenu.js
@@ -1,5 +1,5 @@
 const createProductMenu = (main, document) => {
     const productMenu = main.querySelector('div.product-menu');
-    console.log(productMenu.outerHTML);
+    //console.log(productMenu.outerHTML);
 };
 export default createProductMenu;
